TINY-11740: Move `ClickBehavior` from `api` to `impl`
