+++ modules/sugar/src/main/ts/ephox/sugar/api/dom/Link.ts
@@ -1,7 +1,8 @@
+import * as ClickBehavior from '../../impl/ClickBehavior';
 import { SugarElement } from '../node/SugarElement';
 import * as SugarHead from '../node/SugarHead';
 import * as Attribute from '../properties/Attribute';
-import * as ClickBehavior from './ClickBehavior';
+import * as Platform from '../view/Platform';
 import * as Insert from './Insert';
 
 const addToHead = (doc: SugarElement<Document>, tag: SugarElement<Node>): void => {
@@ -24,8 +25,26 @@ const addStylesheet = (url: string, scope?: SugarElement<Document>): SugarElemen
   return link;
 };
 
-const generateScript = (fn: () => void): string => `<script>(${fn.toString()})()</script>`;
-const getPreventClicksOnLinksScript = (): string => generateScript(ClickBehavior.preventDefaultClickLinkBehavior);
+const generatePreventClicksOnLinksScript = (
+  fn: (isMetaKeyPressed: (e: MouseEvent) => boolean) => void,
+  isMetaKeyPressed: (e: MouseEvent) => boolean
+): string => {
+  return `<script>(${fn.toString()})(${isMetaKeyPressed.toString()})</script>`;
+};
+
+const getPreventClicksOnLinksScript = (): string => {
+  const isMetaKeyPressed = (e: MouseEvent): boolean =>
+    Platform.isMacOS() || Platform.isiOS()
+      ? e.metaKey
+      : e.ctrlKey && !e.altKey;
+
+  const script = generatePreventClicksOnLinksScript(
+    ClickBehavior.preventDefaultClickLinkBehavior,
+    isMetaKeyPressed
+  );
+
+  return script;
+};
 
 export {
   addStylesheet,
